fix: UserGitHubJpaEntity 수정

+++ src/main/java/com/process/clash/application/compete/my/data/GetCompareWithYesterdayData.java
@@ -25,8 +25,8 @@ public class GetCompareWithYesterdayData {
     }
 
     public record ActiveTime(
-            Integer yesterdayActiveTime,
-            Integer todayActiveTime
+            Long yesterdayActiveTime,
+            Long todayActiveTime
     ) {}
 
     public record Contributors(
